+++ Source/WebCore/Modules/networkinfo/NetworkInfoClient.h
@@ -40,7 +40,7 @@ class NetworkInfoClient {
 public:
     virtual ~NetworkInfoClient() { }
 
-    virtual unsigned int bandwidth() const = 0;
+    virtual double bandwidth() const = 0;
     virtual bool metered() const = 0;
    
     virtual void startUpdating() = 0;
